tech-debt: extract seed-testing-theme.helpers (ROK-1183)
